:shirt: Remove Footer component and adjust layout
This commit removes the Footer component from the ‘+layout.svelte’ file, as it was not needed and took up unnecessary space. It also adjusts the width of the section element in the MainView component to fit the screen size better.

The reason for this change is to simplify the layout and improve the appearance of the application. The Footer component was redundant and did not provide any useful information. The width adjustment makes the application more responsive and consistent across different devices.