:sparkles: Add AboutView component
Added a new component `AboutView.svelte` which displays information about the author. The component contains a section with a heading 'About me' and a paragraph showcasing details about the author's current job and university. It's designed to be responsive, adjusting for both desktop and mobile views.

The addition of this component provides an avenue for users or visitors to learn more about the author in a visually appealing manner.