:shower: Remove About page and correct minor syntax issues
The 'About' page (`+page.svelte`) has been removed from the project. Reasons for this removal might be due to redundancy or obsolescence.

In addition to this, minor syntax corrections have been made to both the `rss.xml.ts` and `sitemap.xml.ts` files. These corrections are mainly related to the removal of extra whitespace and the addition of newlines at the end of the files to follow good coding practices.